7seg2ascii spreadsheet
This spreadsheet is used to manipulate the 7 segment codes for the supported ASCII characters of this project.  There is a different mapping for MAIN and SUB.  The Output tab produces a cut-n-paste text sequence for inclusion in the lcd.c source file.  The resulting LUT arrays are used by asc27m() and asc27s(), respectively.